chore: folder refactoring and package updates
