+++ src/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './log'
-export * from './utils'
 export * from './types'
 export * from './vscode'
